Update main.py
EXERCISE_CONFIG Dictionary: I added a list of keywords.

If you do "Dumbbell Curl", the code sees "Curl" and sets the target to 12 reps and increment to 2.5 lbs.

If you do "Squat", it sets the increment to 5 lbs.

If you do "Lateral Raise", it sets the increment to 0 and pushes the rep goal to 15 (because adding weight on lateral raises is very hard).

get_config function: This simple helper checks if your exercise title contains any of our special keywords. If not, it defaults to the standard 5lb / 12 rep rule.